move get_param_access to utilities
