+++ nrpy/helpers/gpu/gpu_kernel.py
@@ -210,22 +210,6 @@ dim3 threads_per_block(threads_in_x_dir, threads_in_y_dir, threads_in_z_dir);"""
         return c_function_call
 
 
-def get_params_access(parallelization: str) -> str:
-    """
-    Return the appropriate params_struct-access prefix for CUDA vs. non-CUDA.
-    E.g. 'd_params[streamid].' vs. 'params->' where 'd_params' is
-    allocated in __constant__ memory rather a pointer passed as a function argument.
-
-    :param parallelization: The parallelization method to use.
-    :returns: The appropriate prefix for accessing the params struct.
-    """
-    if parallelization == "cuda":
-        params_access = "d_params[streamid]."
-    else:
-        params_access = "params->"
-    return params_access
-
-
 if __name__ == "__main__":
     import doctest
     import sys
