+++ nrpy/helpers/gpu/utilities.py
@@ -7,7 +7,23 @@ Authors: Samuel D. Tootle; sdtootle **at** gmail **dot** com
 from typing import Any, Dict, Optional, Tuple
 
 import nrpy.helpers.gpu.cuda_utilities as cuda_utils
-from nrpy.helpers.gpu.gpu_kernel import GPU_Kernel, get_params_access
+from nrpy.helpers.gpu.gpu_kernel import GPU_Kernel
+
+
+def get_params_access(parallelization: str) -> str:
+    """
+    Return the appropriate params_struct-access prefix for CUDA vs. non-CUDA.
+    E.g. 'd_params[streamid].' vs. 'params->' where 'd_params' is
+    allocated in __constant__ memory rather a pointer passed as a function argument.
+
+    :param parallelization: The parallelization method to use.
+    :returns: The appropriate prefix for accessing the params struct.
+    """
+    if parallelization == "cuda":
+        params_access = "d_params[streamid]."
+    else:
+        params_access = "params->"
+    return params_access
 
 
 def generate_kernel_and_launch_code(
